fix: remove falsey build configs
If the `opts.format.includes` call above returns false, a `false` value
is inserted into the array in place of a rollup config object. Each
element of the `createBuildConfigs` is eventually passed into rollup
as an `inputOption`. If one of these elements is `false`, the script will
pass that value as a config to rollup which results in a rollup error
stating: "You must supply an options object to rollup"

This commit fixes that issue by filtering out falsey values.