Drop parse-git-config and use shelljs to find author name through npm or git config
We give npm first preference, if we do not find it there we try git.
If we still do not have a name we try for their email instead on npm first then on git.

Finally if we still dont have a name we prompt the user for the name.